+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -2058,7 +2058,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01M2FPB4SBRKBMMNZWDK0YQ3Y4"
+    id: Optional[str] = "01M2FVS39XQ7GJ961YN8S5K4ZJ"
 
     description: Optional[str] = None