+++ packages/orq-rc/RELEASES.md
@@ -10909,3 +10909,13 @@ Based on:
 - [python v4.15.0-rc.44] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.44] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.44 - packages/orq-rc
+
+## 2026-09-14 11:48:13
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.45] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.45] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.45 - packages/orq-rc